% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsl_data_synthesis.R
\name{bsl_bridge_estimate}
\alias{bsl_bridge_estimate}
\title{Compute a log-mean-exp bridge estimate of the marginal likelihood}
\usage{
bsl_bridge_estimate(loglik)
}
\arguments{
\item{loglik}{Numeric vector of log-likelihood values.}
}
\value{
Scalar log marginal likelihood estimate.
}
\description{
Compute a log-mean-exp bridge estimate of the marginal likelihood
}
