% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsl_data_synthesis.R
\name{bsl_to_mcmc_list_multi}
\alias{bsl_to_mcmc_list_multi}
\title{Convert a multi-chain BSL fit list to coda mcmc.list objects}
\usage{
bsl_to_mcmc_list_multi(fits_multi)
}
\arguments{
\item{fits_multi}{Named list of chain lists (one per model).}
}
\value{
Named list of \code{coda::mcmc.list} objects.
}
\description{
Convert a multi-chain BSL fit list to coda mcmc.list objects
}
