% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_iqd}
\alias{compute_iqd}
\title{Compute Integrated Quadratic Distance (IQD)}
\usage{
compute_iqd(fit, true_params, dist_type, x_grid = NULL)
}
\arguments{
\item{fit}{Stan fit object}

\item{true_params}{List of true parameters}

\item{dist_type}{Distribution type}

\item{x_grid}{Grid of x values for integration}
}
\value{
IQD value
}
\description{
Compute Integrated Quadratic Distance (IQD)
}
