% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_median_bias}
\alias{compute_median_bias}
\title{Compute median bias}
\usage{
compute_median_bias(fit, param_name, true_value)
}
\arguments{
\item{fit}{Stan fit object}

\item{param_name}{Parameter name}

\item{true_value}{True parameter value}
}
\value{
Median bias (median estimate - true value)
}
\description{
Compute median bias
}
