% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_scenario}
\alias{create_scenario}
\title{Create a scenario with specific characteristics}
\usage{
create_scenario(
  scenario_name,
  dist_type,
  n_datasets,
  n_obs_config = c("fixed", "small_var", "large_var", "custom"),
  n_obs_values = NULL,
  summary_config = c("fixed", "mixed_balanced", "mixed_random",
    "mixed_balanced_with_freq", "mixed_random_with_freq", "custom"),
  summary_values = NULL,
  fixed_summary_type = 1,
  fixed_n_obs = 14,
  mu0 = 2,
  tau = 0.4,
  phi = NULL
)
}
\arguments{
\item{scenario_name}{Descriptive name for the scenario}

\item{dist_type}{Distribution type}

\item{n_datasets}{Number of datasets}

\item{n_obs_config}{Configuration for sample sizes: "fixed", "small_var", "large_var", "custom"}

\item{n_obs_values}{Custom vector of sample sizes (if n_obs_config = "custom")}

\item{summary_config}{Configuration for summary types: "fixed", "mixed_balanced", "mixed_random", "custom"}

\item{summary_values}{Custom vector of summary types (if summary_config = "custom")}

\item{fixed_summary_type}{Summary type to use when summary_config = "fixed" (default 1).}

\item{fixed_n_obs}{Sample size to use when n_obs_config = "fixed" (default 14).}

\item{mu0}{Population mean}

\item{tau}{Between-study SD}

\item{phi}{Distribution-specific parameter}
}
\value{
Scenario specification list
}
\description{
Create a scenario with specific characteristics
}
