% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsynth-package.R
\docType{package}
\name{ddsynth-package}
\alias{ddsynth}
\alias{ddsynth-package}
\title{ddsynth: Delay Distribution from Summary Statistics}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Methods to compute delay distributions from summary statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cm401.github.io/ddsynth}
  \item \url{https://github.com/cm401/ddsynth}
  \item Report bugs at \url{https://github.com/cm401/ddsynth/issues}
}

}
\author{
\strong{Maintainer}: Christian Morgenstern \email{c.morgenstern@imperial.ac.uk}

Authors:
\itemize{
  \item Christian Morgenstern \email{c.morgenstern@imperial.ac.uk}
}

}
