% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_model}
\alias{fit_model}
\title{Fit Stan model to simulated data}
\usage{
fit_model(
  sim_data,
  stan_model,
  chains = 4L,
  iter = 10000L,
  warmup = 1000L,
  refresh = 0L,
  control = list(adapt_delta = 0.95, max_treedepth = 12L),
  ...
)
}
\arguments{
\item{sim_data}{Simulated data from generate_hierarchical_data}

\item{stan_model}{Compiled Stan model}

\item{chains}{Number of Markov chains (default 4)}

\item{iter}{Total number of iterations per chain (default 10000)}

\item{warmup}{Number of warmup iterations per chain (default 1000)}

\item{refresh}{How often to print progress (default 0)}

\item{control}{List of control parameters passed to Stan (e.g. adapt_delta)}

\item{...}{Additional arguments to pass to rstan::sampling()}
}
\value{
Stan fit object
}
\description{
Fit Stan model to simulated data
}
