% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_scenario_library}
\alias{generate_scenario_library}
\title{Generate a comprehensive set of scenarios}
\usage{
generate_scenario_library(
  include_homogeneous = TRUE,
  include_mixed = TRUE,
  include_varied_n = TRUE
)
}
\arguments{
\item{include_homogeneous}{Include scenarios with fixed summary types}

\item{include_mixed}{Include scenarios with mixed summary types}

\item{include_varied_n}{Include scenarios with varied sample sizes}
}
\value{
Data frame of scenarios
}
\description{
Generate a comprehensive set of scenarios
}
