% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_model}
\alias{fit_model}
\title{Fit Stan model to simulated data}
\usage{
fit_model(sim_data, stan_model, ...)
}
\arguments{
\item{sim_data}{Simulated data from generate_hierarchical_data}

\item{stan_model}{Compiled Stan model}

\item{...}{Additional arguments to pass to sampling()}
}
\value{
Stan fit object
}
\description{
Fit Stan model to simulated data
}
